ScanResultsStorage: Factor out createHikariDataSource()
Prepare for reusing this code for the file archiver.
