+++ model/build.gradle.kts
@@ -19,6 +19,7 @@
  */
 
 val exposedVersion: String by project
+val hikariVersion: String by project
 val hopliteVersion: String by project
 val jacksonVersion: String by project
 val semverVersion: String by project
@@ -42,6 +43,7 @@ dependencies {
     implementation("com.sksamuel.hoplite:hoplite-core:$hopliteVersion")
     implementation("com.sksamuel.hoplite:hoplite-hocon:$hopliteVersion")
     implementation("com.vdurmont:semver4j:$semverVersion")
+    implementation("com.zaxxer:HikariCP:$hikariVersion")
     implementation("org.jetbrains.exposed:exposed-core:$exposedVersion")
     implementation("org.jetbrains.exposed:exposed-dao:$exposedVersion")
     implementation("org.jetbrains.exposed:exposed-jdbc:$exposedVersion")
